+++ src/structures/case.jl
@@ -18,11 +18,11 @@ for providing the input to a model.
 - **`products::Vector{<:Resource}`** are the resources that should be incorporated into the
   model.
   !!! tip
-      It must containt all [`ResourceEmit`](@ref) in `EnergyModelsBase`, but it is not
+      It must contain all [`ResourceEmit`](@ref)s in `EnergyModelsBase`, but it is not
       necessary that the [`ResourceCarrier`](@ref) are included. It is however advisable to
       include all resources.
 - **`elements::Vector{Vector}`** are the vectors of [`AbstractElement`](@ref)
-  that should be included in the analysis. It must contain at least vectors of nodes and
+  that should be included in the analysis. It must contain at least the vectors of nodes and
   links for an analysis to be useful.
 - **`couplings::Vector{Vector{Function}}`** are the couplings between the individual function
   element types. These elements are represented through a corresponding function, *e.g.*,