+++ test/test_examples.jl
@@ -1,4 +1,10 @@
 @testset "Run examples" begin
+    # Get the global logger and set the loglevel to Warn
+    logger_org = global_logger()
+    logger_new = ConsoleLogger(Warn)
+    global_logger(logger_new)
+
+    # Iterate through all examples and test the examples
     exdir = joinpath(@__DIR__, "../examples")
     files = filter(endswith(".jl"), readdir(exdir))
     for file ∈ files
@@ -10,4 +16,7 @@
         end
     end
     Pkg.activate(@__DIR__)
+
+    # Reset the loglevel
+    global_logger(logger_org)
 end